+++ src/lv1yusha/story.py
@@ -1,0 +1,41 @@
+# -*- coding: utf-8 -*-
+"""Story: lv1 yusha.
+"""
+import os
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
+sys.path.append('storybuilder')
+
+from storybuilder.builder import world as wd
+from . import config as cnf
+
+# episodes
+
+
+# main
+def world():
+    w = wd.World("lv1 yusha project")
+    w.set_db(cnf.CHARAS,
+            cnf.STAGES,
+            cnf.DAYS,
+            cnf.ITEMS,
+            cnf.INFOS,
+            cnf.FLAGS,
+            )
+    return w
+
+
+def story(w: wd.World):
+    return (w.maintitle("レベル１勇者の旅立ち"),
+            )
+
+
+def main(): # pragma: no cover
+    w = world()
+    return w.build(story(w))
+
+
+if __name__ == '__main__':
+    import sys
+    sys.exit(main())
+